Allow for longer response time from MXC network
Discord only lets you respond to an interaction within 3 seconds of it being sent. This means that if it takes longer than that for a command to run (i.e when re-connecting to the MXC network), the interaction will fail. Now I simply respond with a message saying to wait, and edit it when we have a response.